+++ device/scene/surface/material/sampler/ImageSamplerHelpers.cpp
@@ -47,6 +47,48 @@ bool isFloat(ANARIDataType format)
   return false;
 }
 
+bool isFixed16(ANARIDataType format)
+{
+  switch (format) {
+  case ANARI_UFIXED16_VEC4:
+  case ANARI_UFIXED16_VEC3:
+  case ANARI_UFIXED16_VEC2:
+  case ANARI_UFIXED16:
+    return true;
+  default:
+    break;
+  }
+  return false;
+}
+
+bool isFixed8(ANARIDataType format)
+{
+  switch (format) {
+  case ANARI_UFIXED8_VEC4:
+  case ANARI_UFIXED8_VEC3:
+  case ANARI_UFIXED8_VEC2:
+  case ANARI_UFIXED8:
+    return true;
+  default:
+    break;
+  }
+  return false;
+}
+
+bool isSrgb8(ANARIDataType format)
+{
+  switch (format) {
+  case ANARI_UFIXED8_RGBA_SRGB:
+  case ANARI_UFIXED8_RGB_SRGB:
+  case ANARI_UFIXED8_RA_SRGB:
+  case ANARI_UFIXED8_R_SRGB:
+    return true;
+  default:
+    break;
+  }
+  return false;
+}
+
 int numANARIChannels(ANARIDataType format)
 {
   switch (format) {