+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ##### TBD
 - Added support `image1D` sampler subtype
 - Added support for `primitive.index` on `sphere` geometry
+- Added support for missing array types for image samplers
 - Improved object leak warnings on device release
 
 ## 0.4.0